[RSI] Read triples csv (from RSI script) instead of generating them
