+++ internal/sqldb/gcs.go
@@ -1,0 +1,41 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package sqldb
+
+import (
+	"log"
+	"strings"
+)
+
+// Returns bucket name and object prefix string
+func parseGCSPath(gcsPath string) (string, string, bool) {
+	if body, ok := strings.CutPrefix(gcsPath, "gs://"); ok {
+		parts := strings.SplitN(body, "/", 2)
+		if len(parts) == 0 {
+			return "", "", false
+		}
+		bucketName := parts[0]
+		objectPrefix := ""
+		if len(parts) == 2 {
+			objectPrefix = parts[1]
+			if !strings.HasPrefix(objectPrefix, "/") {
+				objectPrefix += "/"
+			}
+		}
+		log.Printf("bucket: %s, prefix: %s", bucketName, objectPrefix)
+		return bucketName, objectPrefix, true
+	}
+	return "", "", false
+}
